+++ ptsa/data/readers/__init__.py
@@ -1,3 +1,5 @@
+import warnings
+
 from .base import *
 from .edf import EDFRawReader
 from .eeg import EEGReader
@@ -10,3 +12,9 @@ from ptsa.data.readers.base import BaseRawReader
 from .tal import *
 from .binary import BinaryRawReader
 from .localization import LocReader
+
+warnings.warn(
+    "PTSA readers will be removed in a future release. Please consider using "
+    "the cmlreaders package instead: https://github.com/pennmem/cmlreaders",
+    FutureWarning
+)