+++ tests/test_filters.py
@@ -13,7 +13,7 @@ from ptsa.data.filters import (
 )
 from ptsa.data.readers import BaseEventReader, EEGReader
 from ptsa.data.readers.tal import TalReader
-from ptsa.test.utils import get_rhino_root, skip_without_rhino
+from tests.utils import get_rhino_root, skip_without_rhino
 
 
 def test_monopolar_to_bipolar_filter_norhino():
@@ -28,15 +28,13 @@ def test_monopolar_to_bipolar_filter_norhino():
         dims=dims, name="test", attrs={'test_attr': 1})
 
     bipolar_pairs1 = np.array([range(9), range(1,10)])
-    m2b1 = MonopolarToBipolarMapper(timeseries=ts,
-                                    bipolar_pairs=bipolar_pairs1)
-    ts_m2b1 = m2b1.filter()
+    m2b1 = MonopolarToBipolarMapper(bipolar_pairs=bipolar_pairs1)
+    ts_m2b1 = m2b1.filter(ts)
 
     bipolar_pairs2 = np.array([(i, j) for i, j in zip(range(9), range(1,10))],
                               dtype=[('ch0', '<i8'), ('ch1', '<i8')])
-    m2b2 = MonopolarToBipolarMapper(timeseries=ts,
-                                    bipolar_pairs=bipolar_pairs2)
-    ts_m2b2 = m2b2.filter()
+    m2b2 = MonopolarToBipolarMapper(bipolar_pairs=bipolar_pairs2)
+    ts_m2b2 = m2b2.filter(ts)
 
     assert np.all(ts_m2b1 == ts_m2b2)
 
@@ -70,9 +68,9 @@ def test_monopolar_to_bipolar_filter_norhino():
     ts2 = timeseries.TimeSeries.create(
         data, rate, coords=coords2,
         dims=dims2, name="test", attrs={'test_attr': 1})
-    m2b3 = MonopolarToBipolarMapper(timeseries=ts2, channels_dim='electrodes',
+    m2b3 = MonopolarToBipolarMapper(channels_dim='electrodes',
                                     bipolar_pairs=bipolar_pairs1)
-    ts_m2b3 = m2b3.filter()
+    ts_m2b3 = m2b3.filter(ts2)
     assert np.all(ts_m2b3.values == ts_m2b1.values)
     # checking each coord is probably redundant (mismatching coords
     # should cause failure in the above assertion), but won't hurt
@@ -91,21 +89,25 @@ def test_monopolar_to_bipolar_filter_norhino():
     assert np.all(ts_m2b3['electrodes'] == bipolar_pairs2)
     assert np.all(ts_m2b3 == (ts.sel(channels=range(9)).values -
                               ts.sel(channels=range(1,10)).values))
-    m2b4 = MonopolarToBipolarMapper(timeseries=ts2, channels_dim='electrodes',
+    m2b4 = MonopolarToBipolarMapper(channels_dim='electrodes',
                                     bipolar_pairs=bipolar_pairs2)
-    ts_m2b4 = m2b4.filter()
+    ts_m2b4 = m2b4.filter(ts2)
     assert np.all(ts_m2b4 == ts_m2b3)
+
     # checking each coord is probably redundant (mismatching coords
     # should cause failure in the above assertion), but won't hurt
     for coord in ts_m2b4.coords:
         assert np.all(ts_m2b3[coord] == ts_m2b4[coord])
+
     # sanity check that we haven't lost any coords:
     for coord in ts.coords:
         if coord != 'channels':
             assert np.all(ts[coord] == ts_m2b4[coord])
+
     for attr in ts.attrs:
         assert np.all(ts_m2b4.attrs[attr] == ts_m2b1.attrs[attr])
         assert np.all(ts.attrs[attr] == ts_m2b4.attrs[attr])
+
     assert ts.name == ts_m2b4.name
     assert np.all(ts_m2b4['electrodes'] == bipolar_pairs2)
     assert np.all(ts_m2b4 == (ts.sel(channels=range(9)).values -
@@ -113,11 +115,12 @@ def test_monopolar_to_bipolar_filter_norhino():
 
     bipolar_pairs3 = np.array([(i, j) for i, j in zip(range(9), range(1,10))],
                               dtype=[('channel0', '<i8'), ('channel1', '<i8')])
-    m2b5 = MonopolarToBipolarMapper(timeseries=ts2, channels_dim='electrodes',
+    m2b5 = MonopolarToBipolarMapper(channels_dim='electrodes',
                                     bipolar_pairs=bipolar_pairs1,
                                     chan_names=['channel0', 'channel1'])
-    ts_m2b5 = m2b5.filter()
+    ts_m2b5 = m2b5.filter(ts2)
     assert np.all(ts_m2b5.values == ts_m2b1.values)
+
     # checking each coord is probably redundant (mismatching coords
     # should cause failure in the above assertion), but won't hurt
     for coord in ts_m2b5.coords:
@@ -126,33 +129,40 @@ def test_monopolar_to_bipolar_filter_norhino():
     for a, b in zip(ts_m2b5['electrodes'], ts_m2b1['channels']):
         assert np.all(
             np.array(a.values.tolist()) == np.array(b.values.tolist()))
+
     # sanity check that we haven't lost any coords:
     for coord in ts.coords:
         if coord != 'channels':
             assert np.all(ts[coord] == ts_m2b5[coord])
+
     for attr in ts.attrs:
         assert np.all(ts_m2b5.attrs[attr] == ts_m2b1.attrs[attr])
         assert np.all(ts.attrs[attr] == ts_m2b5.attrs[attr])
+
     assert ts.name == ts_m2b5.name
     assert np.all(ts_m2b5['electrodes'] == bipolar_pairs3)
     assert np.all(ts_m2b5 == (ts.sel(channels=range(9)).values -
                               ts.sel(channels=range(1,10)).values))
-    m2b6 = MonopolarToBipolarMapper(timeseries=ts2, channels_dim='electrodes',
+    m2b6 = MonopolarToBipolarMapper(channels_dim='electrodes',
                                     chan_names=['channel0', 'channel1'],
                                     bipolar_pairs=bipolar_pairs3)
-    ts_m2b6 = m2b6.filter()
+    ts_m2b6 = m2b6.filter(ts2)
     assert np.all(ts_m2b6 == ts_m2b5)
+
     # checking each coord is probably redundant (mismatching coords
     # should cause failure in the above assertion), but won't hurt
     for coord in ts_m2b6.coords:
         assert np.all(ts_m2b6[coord] == ts_m2b5[coord])
+
     # sanity check that we haven't lost any coords:
     for coord in ts.coords:
         if coord != 'channels':
             assert np.all(ts[coord] == ts_m2b6[coord])
+
     for attr in ts.attrs:
         assert np.all(ts_m2b6.attrs[attr] == ts_m2b1.attrs[attr])
         assert np.all(ts.attrs[attr] == ts_m2b6.attrs[attr])
+
     assert ts.name == ts_m2b6.name
     assert np.all(ts_m2b6['electrodes'] == bipolar_pairs3)
     assert np.all(ts_m2b6 == (ts.sel(channels=range(9)).values -
@@ -203,19 +213,19 @@ class TestFilters(unittest.TestCase):
         session_reader = EEGReader(session_dataroot=dataroot, channels=self.monopolar_channels)
         session_eegs = session_reader.read()
 
-        sedc = DataChopper(events=self.base_events, timeseries=session_eegs,
+        sedc = DataChopper(events=self.base_events,
                            start_time=self.start_time, end_time=self.end_time, buffer_time=self.buffer_time)
-        chopped_session = sedc.filter()
+        chopped_session = sedc.filter(session_eegs)
         assert_array_equal(chopped_session, self.base_eegs)
 
-        sedc = DataChopper(start_offsets=self.base_events.eegoffset, timeseries=session_eegs,
+        sedc = DataChopper(start_offsets=self.base_events.eegoffset,
                            start_time=self.start_time, end_time=self.end_time, buffer_time=self.buffer_time)
-        chopped_session = sedc.filter()
+        chopped_session = sedc.filter(session_eegs)
         assert_array_equal(chopped_session, self.base_eegs)
 
     def test_monopolar_to_bipolar_filter(self):
-        m2b = MonopolarToBipolarMapper(timeseries=self.base_eegs, bipolar_pairs=self.bipolar_pairs)
-        bp_base_eegs = m2b.filter()
+        m2b = MonopolarToBipolarMapper(bipolar_pairs=self.bipolar_pairs)
+        bp_base_eegs = m2b.filter(self.base_eegs)
 
         bipolar_pairs = bp_base_eegs['channels'].data
         for i, bp in enumerate(bipolar_pairs):
@@ -231,16 +241,16 @@ class TestFilters(unittest.TestCase):
         session_reader = EEGReader(session_dataroot=dataroot, channels=self.monopolar_channels)
         session_eegs = session_reader.read()
 
-        m2b = MonopolarToBipolarMapper(timeseries=session_eegs, bipolar_pairs=self.bipolar_pairs)
-        bp_session_eegs = m2b.filter()
+        m2b = MonopolarToBipolarMapper(bipolar_pairs=self.bipolar_pairs)
+        bp_session_eegs = m2b.filter(session_eegs)
 
-        sedc = DataChopper(events=self.base_events, timeseries=bp_session_eegs, start_time=self.start_time,
+        sedc = DataChopper(events=self.base_events, start_time=self.start_time,
                            end_time=self.end_time, buffer_time=self.buffer_time)
 
-        bp_session_eegs_chopped = sedc.filter()
+        bp_session_eegs_chopped = sedc.filter(bp_session_eegs)
 
-        m2b = MonopolarToBipolarMapper(timeseries=self.base_eegs, bipolar_pairs=self.bipolar_pairs)
-        bp_base_eegs = m2b.filter()
+        m2b = MonopolarToBipolarMapper(bipolar_pairs=self.bipolar_pairs)
+        bp_base_eegs = m2b.filter(self.base_eegs)
 
         assert_array_equal(bp_session_eegs_chopped, bp_base_eegs)
 
@@ -248,13 +258,12 @@ class TestFilters(unittest.TestCase):
     @pytest.mark.xfail
     def test_wavelets_with_event_data_chopper(self):
         wf_session = MorletWaveletFilter(
-            timeseries=self.session_eegs[:, :, :int(self.session_eegs.shape[2] / 4)],
             freqs=np.logspace(np.log10(3), np.log10(180), 8),
             output='power',
             verbose=True
         )
 
-        pow_wavelet_session = wf_session.filter()
+        pow_wavelet_session = wf_session.filter(self.session_eegs[:, :, :int(self.session_eegs.shape[2] / 4)])
 
         sedc = DataChopper(events=self.base_events, timeseries=pow_wavelet_session, start_time=self.start_time,
                            end_time=self.end_time, buffer_time=self.buffer_time)
@@ -263,13 +272,13 @@ class TestFilters(unittest.TestCase):
         # removing buffer
         chopped_session_pow_wavelet = chopped_session_pow_wavelet[:, :, :, 500:-500]
 
-        wf = MorletWaveletFilter(timeseries=self.base_eegs,
-                                 freqs=np.logspace(np.log10(3), np.log10(180), 8),
-                                 output='power',
-                                 verbose=True
-                                 )
+        wf = MorletWaveletFilter(
+            freqs=np.logspace(np.log10(3), np.log10(180), 8),
+            output='power',
+            verbose=True
+        )
 
-        pow_wavelet = wf.filter()
+        pow_wavelet = wf.filter(self.base_eegs)
 
         pow_wavelet = pow_wavelet[:, :, :, 500:-500]
 
@@ -283,8 +292,8 @@ class TestFilters(unittest.TestCase):
 
         from xarray.testing import assert_equal
 
-        b_filter = ButterworthFilter(timeseries=self.base_eegs, freq_range=[58., 62.], filt_type='stop', order=4)
-        base_eegs_filtered_1 = b_filter.filter()
+        b_filter = ButterworthFilter(freq_range=[58., 62.], filt_type='stop', order=4)
+        base_eegs_filtered_1 = b_filter.filter(self.base_eegs)
 
         base_eegs_filtered_2 = self.base_eegs.filtered(freq_range=[58., 62.], filt_type='stop', order=4)
 
@@ -298,31 +307,35 @@ class TestFilters(unittest.TestCase):
 def time_series():
     times = np.linspace(0, 1, 1000)
     ts = np.sin(8*times) + np.sin(16*times) + np.sin(32*times)
-    return timeseries.TimeSeries(data=ts, dims=('time'),
-                                            coords={
-                                                'time': times,
-                                                'samplerate': 1000
-                                            })
+    return timeseries.TimeSeries(
+        data=ts,
+        dims=('time'),
+        coords={
+            'time': times,
+            'samplerate': 1000
+        }
+    )
 
 
 @pytest.mark.filters
 class TestFiltersExecute:
     @classmethod
-    def setup_class(cls):
+    def setup_class(self):
         times = np.linspace(0, 1, 1000)
         ts = np.sin(8*times) + np.sin(16*times) + np.sin(32*times)
-        cls.timeseries = timeseries.TimeSeries(data=ts, dims=('time'),
-                                                coords={
-                                                    'time': times,
-                                                    'samplerate': 1000
-                                                })
+        self.timeseries = timeseries.TimeSeries(
+            data=ts, dims=('time'),
+            coords={
+                'time': times,
+                'samplerate': 1000
+            }
+        )
 
-    def test_butterworth(self,time_series):
-        bfilter = ButterworthFilter(timeseries=time_series,
-                                    freq_range=[10., 20.],
+    def test_butterworth(self, time_series):
+        bfilter = ButterworthFilter(freq_range=[10., 20.],
                                     filt_type='stop',
                                     order=2)
-        bfilter.filter()
+        bfilter.filter(time_series)
 
     @pytest.mark.morlet
     @pytest.mark.parametrize('output_type', [
@@ -331,10 +344,9 @@ class TestFiltersExecute:
         ['power', 'phase'],
     ])
     def test_morlet(self, output_type):
-        mwf = MorletWaveletFilter(timeseries=self.timeseries,
-                                  freqs=np.array([10., 20., 40.]),
+        mwf = MorletWaveletFilter(freqs=np.array([10., 20., 40.]),
                                   width=4, output=output_type)
-        result = mwf.filter()
+        result = mwf.filter(self.timeseries)
 
         if len(mwf.output) == 1:
             if 'power' in mwf.output:
@@ -342,11 +354,11 @@ class TestFiltersExecute:
             if 'phase' in mwf.output:
                 assert result.data.shape == (3, 1000)
         else:
-            assert result.data.shape == (2,3,1000)
+            assert result.data.shape == (2, 3, 1000)
 
-    def test_resample(self,time_series):
-        rf = ResampleFilter(timeseries=time_series, resamplerate=50.)
-        new_ts = rf.filter()
+    def test_resample(self, time_series):
+        rf = ResampleFilter(resamplerate=50.)
+        new_ts = rf.filter(timeseries=time_series)
         assert len(new_ts) == 50
         assert new_ts.samplerate == 50.
 
@@ -355,9 +367,9 @@ class TestFilterShapes:
     """Filter behavior should not depend on shape of input array."""
     @classmethod
     def setup_class(self):
-        self.times = times = np.linspace(0,1,1000)
+        self.times = times = np.linspace(0, 1, 1000)
         self.data = np.sin(8*times) + np.sin(16*times) + np.sin(32*times)
-        self.freqs = np.array([10,20],dtype=float)
+        self.freqs = np.array([10, 20], dtype=float)
         self.timeseries = timeseries.TimeSeries(data=self.data[None,:],
                                                 coords={
                                                     'offsets': [0],
@@ -367,13 +379,13 @@ class TestFilterShapes:
                                                 dims=('offsets', 'time'))
 
     def test_morlet(self):
-        results0 = MorletWaveletFilter(self.timeseries, freqs=self.freqs,
-                                       width=4, output=['power','phase']).filter()
+        results0 = MorletWaveletFilter(
+            freqs=self.freqs, width=4, output=['power', 'phase']
+        ).filter(self.timeseries)
 
-        results1 = MorletWaveletFilter(self.timeseries.transpose(),
-                                       freqs=self.freqs,
-                                       width=4, output=['power','phase']).filter()
-        print(results0)
+        results1 = MorletWaveletFilter(
+            freqs=self.freqs, width=4, output=['power', 'phase']
+        ).filter(self.timeseries.transpose())
 
         xr.testing.assert_allclose(results0.sel(output='power'),
                                    results1.sel(output='power'))
@@ -382,8 +394,8 @@ class TestFilterShapes:
 
     @pytest.mark.slow
     def test_butterworth(self):
-        filtered0 = ButterworthFilter(self.timeseries, self.freqs).filter()
-        filtered1 = ButterworthFilter(self.timeseries.transpose(), self.freqs).filter()
+        filtered0 = ButterworthFilter(self.freqs).filter(self.timeseries)
+        filtered1 = ButterworthFilter(self.freqs).filter(self.timeseries.transpose())
 
         xr.testing.assert_allclose(filtered0, filtered1.transpose(*filtered0.dims))
 
@@ -399,33 +411,34 @@ class TestBaseFilter:
                 "samplerate": 1
             }
         )
-
-    @pytest.mark.parametrize("dtype,expected_dtype", [
-        (None, np.float64),
-        (np.float64, np.float64),
-        ("double", np.float64),
-        ("float32", np.float32),
-        ("int", np.int),
-    ])
-    def test_dtypes(self, dtype, expected_dtype):
-        filt = BaseFilter(self.dummy_ts, dtype)
-        assert filt.timeseries.data.dtype == expected_dtype
-
-    @pytest.mark.parametrize("dtype", [np.int8, np.uint16, np.float32])
-    def test_unchanged_dtype(self, dtype):
-        """Special case test where we use a weird input dtype. The previous
-        test works with None since by default Numpy uses float64.
-
-        """
-        ts = self.dummy_ts.astype(dtype)
-        ts = timeseries.TimeSeries(ts.data, ts.coords, ts.dims)
-        filt = BaseFilter(ts, None)
-        assert filt.timeseries.data.dtype == dtype
+# These commented tests don't make sense anymore now that BaseFilter
+# does not accept/apply a dtype
+#    @pytest.mark.parametrize("dtype,expected_dtype", [
+#        (None, np.float64),
+#        (np.float64, np.float64),
+#        ("double", np.float64),
+#        ("float32", np.float32),
+#        ("int", np.int),
+#    ])
+#    def test_dtypes(self, dtype, expected_dtype):
+#        filt = BaseFilter(self.dummy_ts, dtype)
+#        assert filt.timeseries.data.dtype == expected_dtype
+#
+#    @pytest.mark.parametrize("dtype", [np.int8, np.uint16, np.float32])
+#    def test_unchanged_dtype(self, dtype):
+#        """Special case test where we use a weird input dtype. The previous
+#        test works with None since by default Numpy uses float64.
+#
+#        """
+#        ts = self.dummy_ts.astype(dtype)
+#        ts = timeseries.TimeSeries(ts.data, ts.coords, ts.dims)
+#        filt = BaseFilter(ts, None)
+#        assert filt.timeseries.data.dtype == dtype
 
     def test_filter(self):
         with pytest.raises(NotImplementedError):
-            filt = BaseFilter(self.dummy_ts)
-            filt.filter()
+            filt = BaseFilter()
+            filt.filter(self.dummy_ts)
 
 
 class TestMorletFilter:
@@ -447,5 +460,5 @@ class TestMorletFilter:
             }
         )
 
-        mwf = MorletWaveletFilter(ts, np.array(range(70, 171, 10)), output="power")
-        mwf.filter()
+        mwf = MorletWaveletFilter(np.array(range(70, 171, 10)), output="power")
+        mwf.filter(ts)